Significantly Improved Guid Randomness
Its now correctly 64bit instead of 32bit and has more normal distribution.

+++ Engine/GameObject.cpp
@@ -20,7 +20,7 @@ void GameObject::SetState(GameObjectState value) noexcept
 	else OnDeactivate();
 }
 
-unsigned long int GameObject::GetGUID() const noexcept
+unsigned long long int GameObject::GetGUID() const noexcept
 {
 	return guid;
 }
